Move tracing under observability folder
