+++ packages/core/src/errors/index.ts
@@ -10,10 +10,3 @@ export { UnrecognizedCredentialTypeError } from './unrecognized-credential-type.
 export { UnrecognizedNodeTypeError } from './unrecognized-node-type.error';
 
 export { ErrorReporter } from './error-reporter';
-export {
-	Tracing,
-	SpanStatus,
-	type Tracer,
-	type StartSpanOpts,
-	type Span,
-} from './tracing';
