+++ packages/core/src/observability/index.ts
@@ -1,0 +1,8 @@
+export {
+	Tracing,
+	SpanStatus,
+	type Tracer,
+	type StartSpanOpts,
+	type Span,
+} from './tracing/tracing';
+export { SentryTracing } from './tracing/sentry-tracing';
