+++ packages/core/src/observability/tracing/__tests__/sentry-tracing.test.ts
